[cmake] Use LINK_INTERFACE_LIBRARIES to specifiy libraries that a library depends on.
Previously we use to directly link these libraries using
target_link_libraries(${lib} ${other}). This works fine if these are static
libraries. In this case all the symbols of the other library will also be ad
to lib. But problems occur if lib is a shared library and there are static l
in other that were not compiled with -fpic (to be position independent.

The resolve this we now use target_link_libraries(${lib} LINK_INTERFACE_LIBR
which does not incorporate other, but specifies them as needed to linked in
This information is exported by <dune-module>-targets.cmake.

This should solve
https://dune-project.org/flyspray/index.php?do=details&task_id=1429 together
with the corresponding patch from dune-common.