[YaspGrid] Change type of YGrid::Iterator::_which to unsigned from int to avoid a signed/unsigned comparison warning.
A YGrid consists of binom(dim,codim) components, one for each orientation that
entities of this codimension can have.  YGrid::Iterator stores the number of
the component it currently points into in the meber variable _which.  It is
used as an index into wrious vectors in the YGrid and is strictly
non-negative, so this should be safe.