[YaspGrid][YGrid] Implement random-access iterator.
Note that I again had to implement a index.  It might be possible to reuse the
superindex, but that appears to have semantics slightly different from what I
need.  I suspect that we have a special case where index and superindex are
the same, but that depends on whether the "subgrid of an enclosing
grid"-feature of YGridComponent is actually used anywhere.

Note also that it is debatable whether this is actually a random-access
iterator codimensions other than dim and 0.  The reason is that for
codimensions other than dim and 0, there are entities ind multiple
orientations, and each orientation is represented by a YGridComponent.  Each
YGridComponent, even for the same codim, can have a different number of
entities.  When advancing an iterator we have to look up the starting index
for each YGridComponent with a binary search.  Since there are binom(dim,
codim) components, the binary search will introduce a factor of ld(binom(dim,
codim)).  I argue however that this is all part of the constant, since this
factor is independent of the number of elements, and so I will mark all
iterators as random-access.