Remove syntactic sugar and add expectation assertion for some concepts
The methods `is{Concept}` were just a syntactic sugar for the actual
evaluation of the concepts, there was no added value to them. Now, the methods introduced
now `expect{Concept}` assert that a concept is fullfiled. This is not a syntactic sugar
because the implemenation of the concept is forwared to either fallback implementation or
the C++20 concepts. In the first case, if a concept is not fullfiled, the compuler will just
issue an error on the static assert, while on the second, the compiler will indeed describe why
the concept was not fullfiled.